+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ### New features
 
+- Added `\Wizaplace\SDK\Basket\BasketService::setUserBasketId`
+- Added `\Wizaplace\SDK\Basket\BasketService::getUserBasketId`
+
 ### Bugfixes
 
 </details>